# api/docs/swagger.yaml
definitions:
  api.BlocklistsUpdates:
    properties:
      blocklist_ids:
        items:
          type: string
        type: array
    required:
    - blocklist_ids
    type: object
  api.ErrResponse:
    properties:
      details:
        items:
          type: string
        type: array
      error:
        type: string
    type: object
  api.LogoRequest:
    properties:
      domains:
        items:
          type: string
        type: array
    required:
    - domains
    type: object
...
  model.Blocklist:
    properties:
      blocklist_id:
        type: string
      default:
        description: default blocklist is enabled when profile is created
        type: boolean
      description:
        description: displayed to the user
        type: string
      entries:
        type: integer
      homepage:
        type: string
      id:
        type: string
      last_modified:
        type: string
      name:
        description: conventional blocklist name, displayed to the user
        type: string
      source_url:
        type: string
      tags:
        items:
          type: string
        type: array
      type:
        type: string
    required:
    - blocklist_id
    - description
    - name
    type: object
...
  requests.AdvancedOptionsReq:
    properties:
      encryption_type:
        enum:
        - https
        - tls
        type: string
      excluded_wifi_networks:
        description: ExcludedDomains          string `json:"excluded_domains"`
        type: string
    required:
    - encryption_type
    type: object
...
  requests.LoginBody:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  requests.MobileConfigReq:
    properties:
      advanced_options:
        $ref: '#/definitions/requests.AdvancedOptionsReq'
      device_id:
        description: |-
          DeviceId is an optional human-friendly identifier for the device.
          It will be normalized (allowing only [A-Za-z0-9 -]) and truncated to a max length
          consistent with the DNS proxy rules (currently 16). When provided, generated
          mobileconfig profile endpoints (DoH / DoT / DoQ) will embed it so queries can
          be attributed per-device.
        type: string
      profile_id:
        type: string
    required:
    - profile_id
    type: object
...
  responses.ShortLinkResponse:
    properties:
      link:
        type: string
    type: object
...
info:
  contact: {}
  description: modDNS REST API
  title: modDNS REST API
  version: "1.0"
paths:
...
  /api/v1/accounts/logout:
    post:
      description: Logout endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      summary: Logout
      tags:
      - Authentication
...
  /api/v1/auxiliary/logos:
    post:
      consumes:
      - application/json
      description: Download brand logo(s) from Brandfetch. Accepts a list of domains
        and returns a JSON object mapping each domain to its logo as a base64-encoded
        data URL. Errors for each domain are also included.
      parameters:
      - description: Domains to fetch logos for
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LogoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Map of domains to base64-encoded logo data URLs and errors
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      summary: Download brand logo(s) from Brandfetch
      tags:
      - Auxiliary
  /api/v1/blocklists:
    get:
      description: Get available blocklists data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Blocklist'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Get blocklists data
      tags:
      - Blocklists
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Login endpoint
      parameters:
      - description: Login request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.LoginBody'
      - description: MFA OTP code
        in: header
        name: x-mfa-code
        type: string
      - collectionFormat: csv
        description: MFA methods
        in: header
        items:
          type: string
        name: x-mfa-methods
        type: array
      - description: Remove all active sessions before logging in
        in: header
        name: x-sessions-remove
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      summary: Login
      tags:
      - Authentication
  /api/v1/mobileconfig:
    post:
      consumes:
      - application/json
      description: Generate configuration profile for Apple devices
      parameters:
      - description: Generate .mobileconfig request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.MobileConfigReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate configuration profile for Apple devices
      tags:
      - Apple mobileconfig
  /api/v1/mobileconfig/short:
    post:
      consumes:
      - application/json
      description: Generate short link for configuration profile (Apple devices)
      parameters:
      - description: Generate .mobileconfig request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.MobileConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ShortLinkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate short link for configuration profile (Apple devices)
      tags:
      - Apple mobileconfig
...
  /api/v1/profiles/{id}/blocklists:
    delete:
      consumes:
      - application/json
      description: Disable blocklists for a profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: Blocklists to disable
        in: body
        name: blocklist_ids
        required: true
        schema:
          $ref: '#/definitions/api.BlocklistsUpdates'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Disable blocklists
      tags:
      - Profile
    post:
      consumes:
      - application/json
      description: Enable blocklists for a profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: Blocklists to disable
        in: body
        name: blocklist_ids
        required: true
        schema:
          $ref: '#/definitions/api.BlocklistsUpdates'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Enable blocklists
      tags:
      - Profile
...
  /api/v1/short/{code}:
    get:
      description: Download configuration profile for Apple devices from short link
      parameters:
      - description: short code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      summary: Download configuration profile for Apple devices from short link
      tags:
      - Apple mobileconfig
...
swagger: "2.0"